Move sphinx fixtures to conftest.py
