+++ tests/test_sphinx/conftest.py
@@ -32,27 +32,33 @@ parameters available to parse to ``@pytest.mark.sphinx``:
 
 """
 import os
+import pathlib
+import pickle
+import shutil
 
-# import pathlib
-# import shutil
-import re
-
+from docutils.nodes import document
 import pytest
 from sphinx.testing.path import path
 
+SOURCE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "sourcedirs"))
 
+
+# TODO autouse not working, may need to be in root conftest
+# (ideally _build folder should be in tempdir)
 # @pytest.fixture(scope="session", autouse=True)
-# def remove_sphinx_builds():
-#     """ remove all build directories from the test folder
-#     """
-#     srcdirs = pathlib.Path(get_test_source_dir())
-#     for entry in srcdirs.iterdir():  # type: pathlib.Path
-#         if entry.is_dir() and entry.joinpath("_build").exists():
-#             shutil.rmtree(str(entry.joinpath("_build")))
+@pytest.fixture()
+def remove_sphinx_builds():
+    """ remove all build directories from the test folder
+    """
+    yield
+    srcdirs = pathlib.Path(SOURCE_DIR)
+    for entry in srcdirs.iterdir():  # type: pathlib.Path
+        if entry.is_dir() and entry.joinpath("_build").exists():
+            shutil.rmtree(str(entry.joinpath("_build")))
 
 
 @pytest.fixture
-def get_sphinx_app_output():
+def get_sphinx_app_output(file_regression):
     def read(
         app,
         buildername="html",
@@ -60,6 +66,7 @@ def get_sphinx_app_output():
         encoding="utf-8",
         extract_body=False,
         remove_scripts=False,
+        regress_html=False,
     ):
 
         outpath = path(os.path.join(str(app.srcdir), "_build", buildername, filename))
@@ -68,18 +75,43 @@ def get_sphinx_app_output():
 
         content = outpath.text(encoding=encoding)
 
-        if extract_body:
-            body_rgx = re.compile("\\<body\\>(.*)\\</body\\>", re.DOTALL)
-            body_search = body_rgx.search(content)
-            if not body_search:
-                raise IOError("could not find body content of {}".format(path))
-            content = body_search.group(1)
+        if regress_html:
+            # only regress the inner body, since other sections are non-deterministic
+            from bs4 import BeautifulSoup
 
-        if remove_scripts:
-            # remove script environments which can change
-            script_rgx = re.compile("\\<script\\>(.*)\\</script\\>", re.DOTALL)
-            content = script_rgx.sub("<script></script>", content)
+            soup = BeautifulSoup(content, "html.parser")
+            doc_div = soup.findAll("div", {"class": "documentwrapper"})[0]
+            file_regression.check(doc_div.prettify(), extension=".html")
 
         return content
+
+    return read
+
+
+@pytest.fixture
+def get_sphinx_app_doctree(file_regression):
+    def read(
+        app,
+        filename="index.doctree",
+        folder="doctrees",
+        encoding="utf-8",
+        regress=False,
+    ):
+
+        outpath = path(os.path.join(str(app.srcdir), "_build", folder, filename))
+        if not outpath.exists():
+            raise IOError("no output file exists: {}".format(outpath))
+
+        with open(outpath, "rb") as handle:
+            doctree = pickle.load(handle)  # type: document
+
+        # convert absolute filenames
+        for node in doctree.traverse(lambda n: "source" in n):
+            node["source"] = pathlib.Path(node["source"]).name
+
+        if regress:
+            file_regression.check(doctree.pformat(), extension=".xml")
+
+        return doctree
 
     return read
